refactor: make service a validation middleware
