+++ src/user/middleware/index.ts
@@ -1,7 +1,9 @@
+import ServerErrors from '../../errors';
 import type UserModel from '../model';
 import { UserDto } from '../types';
+import { isValidUUID, isValidUserDto } from './helpers';
 
-export default class UserService {
+export default class UserMiddleware {
     constructor(private model: UserModel) {}
 
     async findAll() {
@@ -9,18 +11,30 @@ export default class UserService {
     }
 
     async findOne(id: string) {
+        if (!isValidUUID(id)) {
+            throw ServerErrors.InvalidId;
+        }
+
         return this.model.findOne(id);
     }
 
     async create(userDto: UserDto) {
+        if (!isValidUserDto(userDto)) {
+            throw ServerErrors.InvalidUserBody;
+        }
+
         return this.model.create(userDto);
     }
 
     async update(id: string, userDto: UserDto) {
+        if (!isValidUUID(id)) throw ServerErrors.InvalidId;
+
         return this.model.update(id, userDto);
     }
 
     async delete(id: string) {
+        if (!isValidUUID(id)) throw ServerErrors.InvalidId;
+
         return this.model.delete(id);
     }
 }
